+++ functions/api/profile/[username].js
@@ -1,4 +1,4 @@
-import { getDb } from '../../_utils.js';
+import { getDb } from '../../../shared/utils.js';
 
 export const onRequestGet = async ({ env, params }) => {
   const db = await getDb(env);
